Arreglos varios a la modificacion de alumnos

+++ Utilidades/UtilidadDeFecha.cs
@@ -4,6 +4,6 @@ namespace Trabajo_Práctico_Integrador.Utilidades
 {
     public static class UtilidadDeFecha
     {
-        public static int ObtenerDiferenciaDeFechasEnAnos(DateOnly f1, DateOnly f2) => (ConstanteDeFecha.TIEMPO_CERO + (f1.ToDateTime(TimeOnly.MinValue) - f2.ToDateTime(TimeOnly.MinValue))).Year - 1;
+        public static uint ObtenerDiferenciaDeFechasEnAnos(DateOnly f1, DateOnly f2) => (uint)((ConstanteDeFecha.TIEMPO_CERO + (f1.ToDateTime(TimeOnly.MinValue) - f2.ToDateTime(TimeOnly.MinValue))).Year - 1);
     }
 }
